move patch code into Instrumentor
